update native contract
